% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algae_card.R
\name{algae_card}
\alias{algae_card}
\title{Cards for algae}
\usage{
algae_card(name, data = algae)
}
\arguments{
\item{name}{The vernacular or the scientific name of the algae.}

\item{data}{algae}
}
\value{
A graphical representation of the algae card.
}
\description{
This function generates a pedagogical visual card for a given algae.
Information is accompanied by an image of the species.
}
