% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algae_index_detection.R
\name{Algae_index_detection}
\alias{Algae_index_detection}
\title{Index for Algae detection}
\usage{
Algae_index_detection(
  red_band_path,
  nir_band_path,
  green_band_path,
  ndvi_threshold = 0.1,
  ndwi_threshold = -0.2
)
}
\arguments{
\item{red_band_path}{Expected format: tif or jp2}

\item{nir_band_path}{Expected format: tif or jp2}

\item{green_band_path}{Expected format: tif or jp2}

\item{ndvi_threshold}{From 0.1: low vegetation, works with algae. default: 0.1.}

\item{ndwi_threshold}{With -0.2 the index includes wetlands and aquatic areas. default -0.2.}
}
\value{
A raster containing the binary mask of areas detected as algae.
}
\description{
This function can be used to detect the accumulation of algae in shallow waters or on foreshores.
The NDVI and NDWI indices are combined to extract vegetation (algae) from wetlands or aquatic areas.
}
