% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_algae.R
\name{Distribution_algae}
\alias{Distribution_algae}
\title{Algae map distribution in Europe}
\usage{
Distribution_algae(species, limit = 10000, sample_size = 7000)
}
\arguments{
\item{species}{The scientific name of the algae.}

\item{limit}{Integer. Maximum number of occurrences to download (default: 10000)}

\item{sample_size}{Integer. Maximum number of occurrences to plot (default: 7000)}
}
\value{
A map with geolocalised data on the algae species
}
\description{
This function  generates a spatial distribution map of an algae species in Europe.
The geolocated occurrences come from the GBIF (Global Biodiversity Information Facility) system.
}
