% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_algae.R
\name{info_algae}
\alias{info_algae}
\title{Info on Algae}
\usage{
info_algae(name)
}
\arguments{
\item{name}{The vernacular or the scientific name of the algae.}
}
\value{
A summary message about the algae.
}
\description{
This function provides some information about an algae based on its name.
}
\examples{
info_algae("Chondus crispus")

}
