% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intertidal_zone.R
\name{intertidal_zone}
\alias{intertidal_zone}
\title{Intertidal Zone Visualization}
\usage{
intertidal_zone()
}
\value{
A graphical representation, ggplot object, to better visualize the location of algae on the foreshore.
}
\description{
This function returns a scheme showing the location of algae on the foreshore (intertidal zone).
This stratification is typical of algae in the northeast Atlantic.
}
